requirements.txt
rich
ciscoaxl
flask
flask_simplelogin
zeep
python-dotenv